@@ -15,8 +15,9 @@
         }
     ],
     "require": {
-        "php": ">=5.5.0",
-        "egulias/email-validator": "~2.0"
+        "php": ">=7.0.0",
+        "egulias/email-validator": "~2.0",
+        "dkarlovi/symfony-flex-docker": "^0.1"
     },
     "require-dev": {
         "mockery/mockery": "~0.9.1",
