@@ -1,19 +1,23 @@
 <?php
 
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
+
 /**
  * A base test case with some custom expectations.
  *
  * @author Rouven Weßling
  */
-class SwiftMailerTestCase extends \PHPUnit_Framework_TestCase
+class SwiftMailerTestCase extends \PHPUnit\Framework\TestCase
 {
+    use MockeryPHPUnitIntegration;
+
     public static function regExp($pattern)
     {
         if (!is_string($pattern)) {
             throw PHPUnit_Util_InvalidArgumentHelper::factory(1, 'string');
         }
 
-        return new PHPUnit_Framework_Constraint_PCREMatch($pattern);
+        return new \PHPUnit\Framework\Constraint\RegularExpression($pattern);
     }
 
     public function assertIdenticalBinary($expected, $actual, $message = '')
